requirements.txt
pymongo
python-dotenv
pylint
